src/tests/test_client.py
from sqlitecloud.client import SqliteCloudClient
import os
# Mocking SQCloudConnect and other dependencies would be necessary for more comprehensive testing.


def test_sqlite_cloud_client_init():
    connection_str = os.getenv("TEST_CONNECTION_URL")
    print("Connection:", connection_str)
    client = SqliteCloudClient(connection_str=connection_str)
    assert client is not None

def test_sqlite_cloud_open_and_close_conn():
    connection_str = os.getenv("TEST_CONNECTION_URL")
    print("Connection:", connection_str)
    client = SqliteCloudClient(username='admin',password='5wwqRy9EYU', hostname='hjxijstzm.sqlite.cloud',port=8860)
    conn = client.open_connection()
    assert conn is not None
    client.close_connection(conn)


def test_sqlite_cloud_client_exec_query():
    connection_str = os.getenv("TEST_CONNECTION_URL")
    print("Connection:", connection_str)
    client = SqliteCloudClient(connection_str=connection_str)
    query = "SELECT * FROM your_table_name"
    assert client
    assert query

    # You would need to mock the _open_connection method and SQCloudConnect for more realistic testing.

    # result = client.exec_query(query)
    # assert result
